// generated/orval/fetch/google-health-api/models/civilDateTime.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */
import type { Date } from './date';
import type { TimeOfDay } from './timeOfDay';

/**
 * Civil time representation similar to google.type.DateTime, but ensures that neither the timezone nor the UTC offset can be set to avoid confusion between civil and physical time queries.
 */
export interface CivilDateTime {
  /** Required. Calendar date. */
  date?: Date;
  /** Optional. Time of day. Defaults to the start of the day, at midnight if omitted. */
  time?: TimeOfDay;
}