// generated/orval/fetch/google-health-api/models/rollupDataPoint.ts
export interface RollupDataPoint {
  /** Returned by default when rolling up data points from the `active-energy-burned` data type. */
  activeEnergyBurned?: ActiveEnergyBurnedRollupValue;
  /** Returned by default when rolling up data points from the `active-minutes` data type, or when requested explicitly using the `active-minutes` rollup type identifier. */
  activeMinutes?: ActiveMinutesRollupValue;
  /** Returned by default when rolling up data points from the `active-zone-minutes` data type, or when requested explicitly using the `active-zone-minutes` rollup type identifier. */
  activeZoneMinutes?: ActiveZoneMinutesRollupValue;
  /** Returned by default when rolling up data points from the `activity-level` data type, or when requested explicitly using the `activity-level` rollup type identifier. */
  activityLevel?: ActivityLevelRollupValue;
  /** Returned by default when rolling up data points from the `altitude` data type, or when requested explicitly using the `altitude` rollup type identifier. */
  altitude?: AltitudeRollupValue;
  /** Returned by default when rolling up data points from the `blood-glucose` data type. */
  bloodGlucose?: BloodGlucoseRollupValue;
  /** Returned by default when rolling up data points from the `body-fat` data type, or when requested explicitly using the `body-fat` rollup type identifier. */
  bodyFat?: BodyFatRollupValue;
  /** Returned by default when rolling up data points from the `calories-in-heart-rate-zone` data type, or when requested explicitly using the `calories-in-heart-rate-zone` rollup type identifier. */
  caloriesInHeartRateZone?: CaloriesInHeartRateZoneRollupValue;
  /** Returned by default when rolling up data points from the `core-body-temperature` data type, or when requested explicitly using the `core-body-temperature` rollup type identifier. */
  coreBodyTemperature?: CoreBodyTemperatureRollupValue;
  /** Returned by default when rolling up data points from the `distance` data type, or when requested explicitly using the `distance` rollup type identifier. */
  distance?: DistanceRollupValue;
  /** End time of the window this value aggregates over */
  endTime?: string;
  /** Returned by default when rolling up data points from the `floors` data type, or when requested explicitly using the `floors` rollup type identifier. */
  floors?: FloorsRollupValue;
  /** Returned by default when rolling up data points from the `heart-rate` data type, or when requested explicitly using the `heart-rate` rollup type identifier. */
  heartRate?: HeartRateRollupValue;
  /** Returned by default when rolling up data points from the `hydration-log` data type, or when requested explicitly using the `hydration-log` rollup type identifier. */
  hydrationLog?: HydrationLogRollupValue;
  /** Returned by default when rolling up data points from the `nutrition-log` data type, or when requested explicitly using the `nutrition-log` rollup type identifier. */
  nutritionLog?: NutritionLogRollupValue;
  /** Returned by default when rolling up data points from the `run-vo2-max` data type, or when requested explicitly using the `run-vo2-max` rollup type identifier. */
  runVo2Max?: RunVO2MaxRollupValue;
  /** Returned by default when rolling up data points from the `sedentary-period` data type, or when requested explicitly using the `sedentary-period` rollup type identifier. */
  sedentaryPeriod?: SedentaryPeriodRollupValue;
  /** Start time of the window this value aggregates over */
  startTime?: string;
  /** Returned by default when rolling up data points from the `steps` data type, or when requested explicitly using the `steps` rollup type identifier. */
  steps?: StepsRollupValue;
  /** Returned by default when rolling up data points from the `swim-lengths-data` data type, or when requested explicitly using the `swim-lengths-data` rollup type identifier. */
  swimLengthsData?: SwimLengthsDataRollupValue;
  /** Returned by default when rolling up data points from the `time-in-heart-rate-zone` data type, or when requested explicitly using the `time-in-heart-rate-zone` rollup type identifier. */
  timeInHeartRateZone?: TimeInHeartRateZoneRollupValue;
  /** Returned by default when rolling up data points from the `total-calories` data type, or when requested explicitly using the `total-calories` rollup type identifier. */
  totalCalories?: TotalCaloriesRollupValue;
  /** Returned by default when rolling up data points from the `weight` data type, or when requested explicitly using the `weight` rollup type identifier. */
  weight?: WeightRollupValue;
}